requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
selenium==4.15.2
pytest==7.4.3
flask==3.0.0
flask-cors==4.0.0
undetected-chromedriver==3.5.5
